refactor: More rework from feedback
